+++ packages/blogpress/CHANGELOG.md
@@ -4,7 +4,12 @@ publish: false
 
 # CHANGELOG
 
+## 2.0.52
+### Patch Changes
+- @sugarat/theme@0.4.4
+
 ## 2.0.51
+
 ### Patch Changes
 
 - Updated dependencies